new lib created for fortune presentation
