+++ site/lib/fortune.js
@@ -1,0 +1,12 @@
+var fortunes = [
+    "Conquer your fears or they will conquer you.",
+    "Rivers need springs.",
+    "Do not fear what you don't know.",
+    "You will have a pleasant surprise.",
+    "Whenever possible, keep it simple.",
+    ];
+
+exports.getFortune = ()=>{
+    var idx = fortunes[Math.floor(Math.random() * fortunes.length)];
+    return fortunes[idx];
+}
