+++ site/meadowlark.js
@@ -1,4 +1,5 @@
 var express = require('express');
+var fortune = require('./lib/fortune.js');
 
 var app = express();
 
@@ -7,13 +8,7 @@ var handlebars = require('express3-handlebars').create({defaultLayout:'main'});
 app.engine('handlebars', handlebars.engine);
 app.set('view engine', 'handlebars');
 
-var fortunes = [
-    "Conquer your fears or they will conquer you.",
-    "Rivers need springs.",
-    "Do not fear what you don't know.",
-    "You will have a pleasant surprise.",
-    "Whenever possible, keep it simple.",
-    ];
+
 
 
 app.set('port', process.env.PORT || 3000);
@@ -25,8 +20,8 @@ app.get('/', (req, res)=>{
 });
 
 app.get('/about', (req, res)=>{
-    var randomFortune = fortunes[Math.floor(Math.random() * fortunes.length)];
-    res.render('about',{fortune: randomFortune});
+    
+    res.render('about',{fortune: fortune.getFortune});
 });
